Changed the way partitions are created by dbscan.
The unclassified cluster is now removed from the partition before dbscan returns, so that we
don't get a cluster that's always empty in the output.  cdbw has been modified to work with this
new density representation.

Would be really nice to make cdbw a little more general -- it should work on more than just dbscan
clusterings.  The right way to do this would be to allow removing clusters from partitions,
which is currently cumbersome b/c there's no way to have an object that is not in a cluster.
Consider changing type of cluster ids in the future so that they can be negative and algorithms
and (like partitions) can have special clusters ids for things like noise and non-member objects.