Changed types of medoid and object ids to signed.
This allows clustering algorithms to have special internal cluster ids and for objects in partitions to be unclassified.  This also allows clusters to be removed from partitions, so that analyses like cdbw may only apply to certain clusters in data sets.

This change reworks the ids, and it reworks cdbw so that it is independent of density clustering and can be applied to general partitions (not just those with noise clusters).